passport

+++ resolvers.js
@@ -1,8 +1,28 @@
 import models from './model.js';
+import {AuthenticationError} from 'apollo-server-express';
+import {login} from './passport/authController.js';
 
 const Resolvers = {
     Query: {
-        stations: (parent, args) => {
+        login: async (parent, args, {req, res}) => {
+            try {
+                req.body = args
+                const result = await login(req, res);
+                console.log("login result", result);
+                return result;
+            } catch (e) {
+                console.log(e);
+                throw new AuthenticationError("Login failed");
+            }
+        },
+
+        stations: (parent, args, context) => {
+
+            console.log("user is", context.user)
+            if (!context.user) {
+                throw new AuthenticationError("Requires logged in user")
+            }
+
             const limit = args.limit || 10
             const {topRight, bottomLeft} = args
 
@@ -34,7 +54,13 @@ const Resolvers = {
                 return models.stations.find().populate("Connections").limit(limit).exec()
             }
         },
-        station: (parent, args) => {
+        station: (parent, args, context) => {
+
+            console.log("user is", context.user)
+            if (!context.user) {
+                throw new AuthenticationError("Requires logged in user")
+            }
+
             return models.stations
                 .findById(args.id)
                 .populate("Connections")
@@ -43,16 +69,33 @@ const Resolvers = {
     },
 
     Mutation: {
-        removeStation: (parent, args) => {
+        removeStation: (parent, args, context) => {
+
+            console.log("user is", context.user)
+            if (!context.user) {
+                throw new AuthenticationError("Requires logged in user")
+            }
+
             return models.stations.deleteOne({_id: args.id});
         },
 
-        modifyStation: (parent, args) => {
+        modifyStation: (parent, args, context) => {
+
+            console.log("user is", context.user)
+            if (!context.user) {
+                throw new AuthenticationError("Requires logged in user")
+            }
+
             const argsWithoutId = {...args, id: undefined}
             return models.stations.updateOne({_id: args.id}, argsWithoutId);
         },
 
-        addStation: async (parent, args) => {
+        addStation: async (parent, args, context) => {
+
+            console.log("user is", context.user)
+            if (!context.user) {
+                throw new AuthenticationError("Requires logged in user")
+            }
 
             const connections = args.Connections || []
